Shared Axis Transition (#58)
* Implement shared x- and y-axis transitions

* Incorporate shared z-axis transition into the same builder as shared x- and y-axis transitions
