+++ packages/animations/lib/animations.dart
@@ -5,4 +5,4 @@
 export 'src/fade_through_transition.dart';
 export 'src/open_container.dart';
 export 'src/page_transition_switcher.dart';
-export 'src/shared_z_axis_transition.dart';
+export 'src/shared_axis_transition.dart';
