+++ packages/animations/pubspec.yaml
@@ -13,3 +13,4 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  vector_math: ^2.0.8
